重构代码并创建单独的路径页面
